backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
python-multipart==0.0.6
google-cloud-storage==2.8.0
google-cloud-pubsub==2.17.0
pdf2text==1.0.0
python-dotenv==1.0.0
pydantic==1.10.7
pytest==7.3.1
locust==2.15.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyPDF2==3.0.1 